* test files renamed for vows
