+++ lib/lint/formatter/xml.js
@@ -59,9 +59,9 @@ Formatter.prototype._formatNormal = function (data) {
 	
 			output += this._line(1, '<file name="' + file + '">');
 			output += this._line(2, '<issue'
+					+ ' ' + this._attribute('line', error.line)
 					+ ' ' + this._attribute('char', error.character)
 					+ ' ' + this._attribute('evidence', error.evidence)
-					+ ' ' + this._attribute('line', error.line)
 					+ ' ' + this._attribute('reason', error.reason)
 					+ ' />'
 			);
@@ -98,7 +98,7 @@ Formatter.prototype._attribute = function (name, value) {
 };
 
 Formatter.prototype._escape = function (string) {
-	return (string) ? string.replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;') : '';
+	return (string) ? ('' + string).replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;') : '';
 };
 
 /**
