+++ test/lint/formatter/xml_test.js
@@ -11,9 +11,9 @@ function createReport() {
 		{
 			file : 'foo',
 			error: {
-				character: 'c',
+				line: 12,
+				character: 5,
 				evidence: 'e',
-				line: 'l',
 				reason: 'r'
 			}
 		}
@@ -29,7 +29,7 @@ var FormatterTest = vows.describe('Formatter class').addBatch({
 			assert.equal(topic.format([]), '<?xml version="1.0" encoding="UTF-8" ?><jslint></jslint>');
 		},
 		'should return format simple content' : function (topic) {
-			var expected =  '<?xml version="1.0" encoding="UTF-8" ?><jslint><file name="foo"><issue char="c" evidence="e" line="l" reason="r" /></file></jslint>';
+			var expected =  '<?xml version="1.0" encoding="UTF-8" ?><jslint><file name="foo"><issue line="12" char="5" evidence="e" reason="r" /></file></jslint>';
 			
 			assert.equal(topic.format(createReport()), expected);
 		}
@@ -43,7 +43,7 @@ var FormatterTest = vows.describe('Formatter class').addBatch({
 			assert.equal(topic.format([]), expected);
 		},
 		'should return pretty xml formatted content' : function (topic) {
-			var expected = '<?xml version="1.0" encoding="UTF-8" ?>\n<jslint>\n\t<file name="foo">\n\t\t<issue char="c" evidence="e" line="l" reason="r" />\n\t</file>\n</jslint>\n';
+			var expected = '<?xml version="1.0" encoding="UTF-8" ?>\n<jslint>\n\t<file name="foo">\n\t\t<issue line="12" char="5" evidence="e" reason="r" />\n\t</file>\n</jslint>\n';
 			assert.equal(topic.format(createReport()), expected);
 		}
 	},
@@ -56,7 +56,7 @@ var FormatterTest = vows.describe('Formatter class').addBatch({
 			assert.equal(topic.format([]), expected);
 		},
 		'should return pretty xml formatted content' : function (topic) {
-			var expected = '<?xml version="1.0" encoding="UTF-8" ?>\n<jslint>\n  <file name="foo">\n    <issue char="c" evidence="e" line="l" reason="r" />\n  </file>\n</jslint>\n';
+			var expected = '<?xml version="1.0" encoding="UTF-8" ?>\n<jslint>\n  <file name="foo">\n    <issue line="12" char="5" evidence="e" reason="r" />\n  </file>\n</jslint>\n';
 			assert.equal(topic.format(createReport()), expected);
 		}
 	},
@@ -69,7 +69,7 @@ var FormatterTest = vows.describe('Formatter class').addBatch({
 			assert.equal(topic.format([]), expected);
 		},
 		'should return pretty xml formatted content' : function (topic) {
-			var expected = '<?xml version="1.0" encoding="UTF-8" ?>\n\r<jslint>\n\r\t<file name="foo">\n\r\t\t<issue char="c" evidence="e" line="l" reason="r" />\n\r\t</file>\n\r</jslint>\n\r';
+			var expected = '<?xml version="1.0" encoding="UTF-8" ?>\n\r<jslint>\n\r\t<file name="foo">\n\r\t\t<issue line="12" char="5" evidence="e" reason="r" />\n\r\t</file>\n\r</jslint>\n\r';
 			assert.equal(topic.format(createReport()), expected);
 		}
 	}
