+++ test/lint/formatter_test.js
